General fixes & new build directories
[Change] VulkanEngine/VulkanRendererWindows/VulkanGUI.vcxproj, OutDir is now "$(SolutionDir)$(Platform)\$(Configuration)\"
[Change] VulkanEngine IntDir is now "$(SolutionDir)$(Platform)\$(Configuration)\BuildFiles\"
[Change] VulkanRendererWindows, "$(SolutionDir)$(Platform)\$(Configuration)\BuildFiles\Renderer"
[Change] VulkanGUI.vcxproj, "$(SolutionDir)$(Platform)\$(Configuration)\BuildFiles\GUI"

[Fix] GUITypes.cpp, ITextSlider struct, DrawElement function, static_cast now to int instead of size_t.

+++ VulkanGUI/GUI/GUITypes.cpp
@@ -97,7 +97,7 @@ namespace GUI
 			{
 				return false;
 			}
-			return ImGui::SliderInt(texts[currentValue], &currentValue, 0, static_cast<size_t>(texts.size() - 1));
+			return ImGui::SliderInt(texts[currentValue], &currentValue, 0, static_cast<int>(texts.size() - 1));
 		}
 
 		bool FSlider::DrawElement()
